pyproject.toml
[project]
name = "zendesk-rag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.14.3",
    "google-cloud-storage>=3.8.0",
    "lxml>=6.0.2",
    "markdownify>=1.2.2",
    "playwright>=1.57.0",
    "requests>=2.32.5",
]
